[usergen] Do not remove and re-add already existing users.
The logic is:
1. Generate name-id pairs required by droid layer to droid.ids.
2. On installation create all the required user and groups and note their
   names and ids to droid.ids.created.
3. On update cross-check the two files and only add and remove the name-ids
   that have been changed.

ad. 1. usergroupgen.c no longer creates shellscript output, instead it creates
droid.ids which is a simple text file with name-space-id format, one per line.

ad. 2. postinstall script writes down every name-id it managed to create. It
does not write any user creation of which failed.

ad. 3. it is not verified during user/group removal if the user/group was
modified by the user or with a third-party script. If the entry disappeared
from droid-hal, we just remove it.

Special note: A conversion mechanism was added for upgrades from previous dhd
releases in the preinstall script. All users that were added in the previous
release are considered to have been created.

Change-Id: I2a13462bed5297f3c4ef601fcd26194c59444a3f